+++ leadnews-article/src/main/java/com/tanran/api/ArticleJarApplication.java
@@ -1,23 +1,22 @@
-package com.tanran.user;
+package com.tanran.api;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+
 
 /**
  * TODO
  *
  * @author liuht26262@yunrong.cn
  * @version V3.0
- * @description
- * @since 3.0.x 2022/3/20 21:23
+ * @since 3.0.x 2022/3/18 9:52
  */
+
 @SpringBootApplication
 @MapperScan("com.tanran.model.mappers.app")
-public class UserJarApplication {
+public class ArticleJarApplication {
     public static void main(String[] args) {
-        SpringApplication.run(UserJarApplication.class,args);
+        SpringApplication.run(ArticleJarApplication.class,args);
     }
 }
