+++ leadnews-article/src/main/java/com/tanran/api/service/ArticleContentService.java
@@ -1,0 +1,16 @@
+package com.tanran.api.service;
+
+import com.tanran.common.result.RespResult;
+
+/**
+ * TODO
+ *
+ * @author liuht26262@yunrong.cn
+ * @version V3.0
+ * @description
+ * @since 3.0.x 2022/3/19 22:15
+ */
+
+public interface ArticleContentService {
+    RespResult getArticleContent(Integer articleId);
+}
