--- leadnews-article/src/main/java/com/tanran/article/config/MysqlConfig.java
@@ -1,17 +1,0 @@
-package com.tanran.article.config;
-
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * TODO
- *
- * @author liuht26262@yunrong.cn
- * @version V3.0
- * @since 3.0.x 2022/3/18 10:14
- */
-@Configuration
-@ComponentScan("com.tanran.common.mysql.core")
-public class MysqlConfig {
-}
