+++ leadnews-common/maven_prod.properties
@@ -9,8 +9,8 @@ mysql.core.jdbc.username=root
 mysql.core.jdbc.password=toor
 mysql.core.jdbc.driver=com.mysql.jdbc.Driver
 mysql.core.root.mapper=mappers
-mysql.core.aliases.package=com.heima.**.mysql.core.model.pojos.**
-mysql.core.tx.scan.package=execution(* com.heima..service.*.*(..))
+mysql.core.aliases.package=com.tanran.**.mysql.core.model.pojos.**
+mysql.core.tx.scan.package=execution(* com.tanran..service.*.*(..))
 
 # admin 账户名
 admin.username=admin
