+++ leadnews-common/src/main/java/com/tanran/common/jackson/ConfusionModule.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 public class ConfusionModule extends Module {
 
     public final static String MODULE_NAME = "jackson-confusion-encryption";
-    public final static Version VERSION = new Version(1,0,0,null,"heima",MODULE_NAME);
+    public final static Version VERSION = new Version(1,0,0,null,"tanran",MODULE_NAME);
 
     @Override
     public String getModuleName() {
