+++ leadnews-login/maven_dev.properties
@@ -1,0 +1,1 @@
+profiles.name=dev
