+++ leadnews-login/maven_prod.properties
@@ -1,0 +1,1 @@
+profiles.name=prod
