+++ leadnews-login/maven_test.properties
@@ -1,0 +1,3 @@
+# log4j
+log.level=DEBUG
+log.pattern=%d{DEFAULT}^|%sn^|%level^|%t^|%c^|%M^|%msg%n
