+++ leadnews-login/src/main/resources/application.yml
@@ -1,0 +1,5 @@
+server:
+  port: ${port.login}
+spring:
+  application:
+    name: ${sn.login}
