+++ leadnews-model/src/main/java/com/tanran/model/mappers/app/ApUserMapper.java
@@ -1,5 +1,8 @@
 package com.tanran.model.mappers.app;
 
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
 import com.tanran.model.user.pojos.ApUser;
 
 /**
@@ -10,6 +13,8 @@ import com.tanran.model.user.pojos.ApUser;
  * @description
  * @since 3.0.x 2022/3/23 10:51
  */
+
 public interface ApUserMapper {
-    public ApUser selectUserById(Long id);
+    ApUser selectUserById( Long id);
+    ApUser selectUserByPhone( String phone);
 }
