+++ leadnews-model/src/main/resources/mappers/app/ApUserMapper.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.tanran.model.mappers.app.ApUserMapper">
     <resultMap id="BaseResultMap" type="com.tanran.model.user.pojos.ApUser" >
         <id column="id" property="id" />
@@ -21,4 +21,11 @@
     <select id="selectUserById" resultMap="BaseResultMap">
         select  <include refid="Base_Column_List" /> from ap_user where id = #{id}
     </select>
+
+    <select id="selectUserByPhone" resultMap="BaseResultMap">
+        select
+        <include refid="Base_Column_List" />
+        from ap_user where phone = #{phone} limit 1;
+    </select>
+
 </mapper>
