+++ leadnews-mybatisGenertator/src/main/resources/generatorConfig.xml
@@ -43,7 +43,7 @@
         <javaClientGenerator targetPackage="com.tanran.model.mappers.app" targetProject="src/main/java" type="XMLMAPPER"></javaClientGenerator>
 
         <!-- table标签可以有多个，至少一个，tableName指定表名，可以使用_和%通配符 -->
-        <table tableName="ap_read_behavior">
+        <table tableName="ap_user">
             <!-- 是否只生成POJO对象 -->
             <property name="modelOnly" value="false"/>
             <!-- 数据库中表名有时我们都会带个前缀，而实体又不想带前缀，这个配置可以把实体的前缀去掉 -->
