+++ leadnews-user/src/main/java/com/tanran/user/controller/UserRelationController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.tanran.article.userApi.UserRelationApi;
+import com.tanran.api.userApi.UserRelationApi;
 import com.tanran.common.result.RespResult;
 import com.tanran.model.user.dtos.UserRelationDto;
 import com.tanran.user.service.UserRelationService;
